refactored to include appearance settings keys
